% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{EncryptAndSave}
\alias{EncryptAndSave}
\title{Utility to encrypt a clear text with a public key
from previously generated cryptographic key files}
\usage{
EncryptAndSave(clear.pw, pub.k, pwRoot = "SSpw")
}
\arguments{
\item{clear.pw}{- the clear text}

\item{pub.k}{- a public encrypted key}

\item{pwRoot}{='SSpw'- the root you want for the encrypted file name}
}
\value{
the encrypted.pw.filename
}
\description{
Utility to encrypt a clear text with a public key
from previously generated cryptographic key files
}
\examples{
\dontrun{
Fill these in with the timestamped file names
RSAPrvKeyFilename<-my.keys$RSAPrvKeyFilename
RSAPubKeyFilename<-my.keys$RSAPubKeyFilename


key.ring<-LoadKeys(RSAPrvKeyFilename,RSAPubKeyFilename)

clear.pw<-"A secret password"

encrypted.pw.filename<-EncryptAndSave(clear.pw,key.ring$pub.k)
}

}
