% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{ErrorBar_LL_UL}
\alias{ErrorBar_LL_UL}
\title{Utility to provide Lower and Upper Limit values for Error Bars.}
\usage{
ErrorBar_LL_UL(x, type = "SD", CI = 0.95)
}
\arguments{
\item{x}{- the data}

\item{type}{- either SD, SE, or CI}

\item{CI}{- confidence interval, defaults to 95\%}
}
\value{
results with ymin and ymax
}
\description{
Utility to provide Lower and Upper Limit values for Error Bars.
}
\examples{
\dontrun{
library(ggplot)
p<-ggplot(data=plot.data,aes(x=measured.value,y=outcome))
p<-p+geom_point(size=2,col='blue')
p<-p+stat_summary(fun.y='mean', geom="point", size = 5, shape=15,col='red')
This style is from the old ggplot v1 code
p<-p+stat_summary(fun.data = 'ErrorBar_LL_UL',type='CI',CI=0.95, geom = "errorbar", width = 10)
This style is from the newer ggplot v2 code ... hopefully you are at this level or above
p<-p+stat_summary(fun.data = 'ErrorBar_LL_UL',fun.args=list(type='CI',CI=0.95), geom = "errorbar", width = 10)
p
}
}
