% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{FormatJSON_String}
\alias{FormatJSON_String}
\title{Utility to return a formatted JSON string}
\usage{
FormatJSON_String(json.text, filename = NULL)
}
\arguments{
\item{json.text}{- the JSON string}

\item{filename}{= NULL, if you provide a string it will be written to that file
in the current folder. Provide a full path if you want it to go somewhere else}
}
\value{
a formatted version of the json.text
}
\description{
Utility to return a formatted JSON string
}
\examples{
\dontrun{
json.text<-SomeJSON_string_from_SomeWhere
FormatJSON_String(json.text)
FormatJSON_String(json.text,filename='MyJSON_File.json')
}

}
