% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{GenAndSaveKeys}
\alias{GenAndSaveKeys}
\title{Utility to generate encryption keys}
\usage{
GenAndSaveKeys()
}
\value{
list with two file names to private and public key files (.Rdata format)
}
\description{
Utility to generate encryption keys
}
\examples{
\dontrun{
my.keys<-GenAndSaveKeys()
}

}

