% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{GetColourBlindPalette}
\alias{GetColourBlindPalette}
\title{Utility to return a colour blind sensitive colour palette}
\usage{
GetColourBlindPalette(size = NULL)
}
\arguments{
\item{size}{= NULL - provide a size to enable colour recycling}
}
\value{
a character vector with the colour blind palette codes
}
\description{
Utility to return a colour blind sensitive colour palette
}
\examples{
\dontrun{
cbp<-GetColourBlindPalette()
cbp<-GetColourBlindPalette(size=16)
cbp<-GetColourBlindPalette(size=33)
}

}

