% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{GetPaletteSize}
\alias{GetPaletteSize}
\title{Utility to return the palette size for GetColourBlindPalette()}
\usage{
GetPaletteSize(num.groups, palette.block.size = 8)
}
\arguments{
\item{num.groups}{- provide a size to enable colour recycling}

\item{palette.block.size}{= 8 - size of the block of colours used by GetColourBlindPalette()}
}
\value{
the palette size - this is a multiple of palette.block.size
}
\description{
Utility to return the palette size for GetColourBlindPalette()
}
\examples{
\dontrun{
num.process.runs <- 22
palette.size <- GetPaletteSize(num.process.runs)
cbp<-GetColourBlindPalette(size=palette.size)
}

}

