% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{LoadKeys}
\alias{LoadKeys}
\title{Utility to load previously generated cryptographic key files}
\usage{
LoadKeys(RSAPrvKeyFilename, RSAPubKeyFilename)
}
\arguments{
\item{RSAPrvKeyFilename}{- a file name}

\item{RSAPubKeyFilename}{- a file name}
}
\value{
a list with two encrypted key objects
}
\description{
Utility to load previously generated cryptographic key files
}
\examples{
\dontrun{
Fill these in with the timestamped file names
RSAPrvKeyFilename<-my.keys$RSAPrvKeyFilename
RSAPubKeyFilename<-my.keys$RSAPubKeyFilename


key.ring<-LoadKeys(RSAPrvKeyFilename,RSAPubKeyFilename)
private.key<-key.ring$priv.k
public.key<-key.ring$pub.k
}

}

