% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{MakePie}
\alias{MakePie}
\title{Utility to make a ggplot pie chart}
\usage{
MakePie(type, percent, type.title.str = "Title String Here",
  segment.name = "Type", segment.cols = NULL, segment.breaks = NULL,
  label.size = 4)
}
\arguments{
\item{type}{- a character vector of pie segments names}

\item{percent}{- a numeric vector of pie segments values (make them add to 100)}

\item{type.title.str}{= 'Title String Here' - a title for the plot}

\item{segment.name}{= 'Type' - the legend title}

\item{segment.cols}{= NULL - used to control colours of the segments}

\item{segment.breaks}{= NULL - used to control colours of the segments}

\item{label.size}{= 4 - used to adjust percent text label size}
}
\value{
a ggplot object with your pie chart
}
\description{
Utility to make a ggplot pie chart
}
\examples{
\dontrun{
# Get these two vectors in the order you need and make sure
# the percent add up to 100
type<-c('Comp1','Comp2','Comp3','Other')
percent<-c(65,20,10,5)

#Set up the pie chart characteristics
pie.trace.cols<-c('Comp1'='cornflowerblue','Comp2'='forestgreen','Comp3'='orange','Other'='indianred2')
pie.breaks<-c('Comp1','Comp2','Comp3','Other')
pie.segment.name<-'Variation Type'
type.title.str<-'Three Variation Components'

p<-MakePie(type,percent,type.title.str
           ,segment.name=pie.segment.name
           ,segment.cols=pie.trace.cols
           ,segment.breaks=pie.breaks)
print(p)
}

}
