% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{MD5_thisFile}
\alias{MD5_thisFile}
\title{Utility to return MD5 signature of a file}
\usage{
MD5_thisFile(filename, writeMD5File = TRUE)
}
\value{
an uppercase MD5 signature
}
\description{
Utility to return MD5 signature of a file
}
\examples{
\dontrun{
MD5_thisFile(filename)
}

}

