% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{OutputPlotForPaper}
\alias{OutputPlotForPaper}
\title{Wrapper function for plotting ggplot2 graphics}
\usage{
OutputPlotForPaper(plot.folder = "./", filename, plot.obj = NULL,
  plot.width = NULL, plot.height = NULL, units = "in", res = 600)
}
\arguments{
\item{plot.folder}{- where to put the plot}

\item{plot.obj}{- a ggplot2 object}

\item{plot.filename}{- the filename for the plot}

\item{plot.width=NULL}{- used to specify width}

\item{plot.height=NULL}{- used to specify height}

\item{units='in'}{- inches}

\item{res=72}{- this gives a small file by default}
}
\description{
Wrapper function for plotting ggplot2 graphics
}
\examples{
\dontrun{

x.vals<-c(1,2,3)
y.vals<-c(2,4,6)
plot.data<-data_frame(x.vals,y.vals)
p<-ggplot(data = plot.data,aes(x=x.vals,y=y.vals))
p<-p+geom_line(size=0.5)
p<-p+geom_point(size=2)
titleStr<-'A simple plot'
p<-p+labs(title = titleStr)

print(p)

file.name<-'A good file name.pdf'
OutputPlotForPaper(file.name,plot.obj=p)
}
}
