% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{PlotPDF}
\alias{PlotPDF}
\title{Utility to wrap the standard pdf() driver to A4 or A3 format}
\usage{
PlotPDF(plot.filename, page.size = "A4", landscape = TRUE)
}
\arguments{
\item{plot.filename}{- the filename for the plot}

\item{page.size='A4'}{- used to specify page size, can be A3}

\item{landscape=TRUE}{- used to specify page orientation}
}
\description{
Remember to use dev.off() after you have done the plot.
}
\examples{
\dontrun{
filename<-'A_good_plot.pdf'
PlotPDF(filename)
 ... some plot commands
dev.off()

PlotPDF(filename,page.size='A3',landscape=FALSE)
 ... some plot commands
dev.off()
}
}

