% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{ProducePlot}
\alias{ProducePlot}
\title{Wrapper function for plotting ggplot2 PDFs}
\usage{
ProducePlot(filename, plot.obj, page.size = "A4", landscape = TRUE)
}
\arguments{
\item{plot.obj}{- a ggplot2 object}

\item{plot.filename}{- the filename for the plot}

\item{page.size='A4'}{- used to specify page size, can be A3}

\item{landscape=TRUE}{- used to specify page orientation}
}
\description{
Wrapper function for plotting ggplot2 PDFs
}
\examples{
\dontrun{

x.vals<-c(1,2,3)
y.vals<-c(2,4,6)
plot.data<-data_frame(x.vals,y.vals)
p<-ggplot(data = plot.data,aes(x=x.vals,y=y.vals))
p<-p+geom_line(size=0.5)
p<-p+geom_point(size=2)
titleStr<-'A simple plot'
p<-p+labs(title = titleStr)

print(p)

file.name<-'A good file name.pdf'
ProducePlot(file.name,plot.obj=p)
}
}

