% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{ProgressDot}
\alias{ProgressDot}
\title{A text based progress bar}
\usage{
ProgressDot(count, total)
}
\arguments{
\item{count}{- the current count}

\item{total}{- the final count}
}
\description{
A text based progress bar
}
\examples{
\dontrun{
ProgressDot(count,my.total.loop.size)
}
}

