% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\docType{package}
\name{RobsRUtils}
\alias{RobsRUtils}
\alias{RobsRUtils-package}
\title{RobsRUtils: Routines I have found or created along the way}
\description{
RobsRUtils: Routines I have found or created along the way
}
\section{RobsRUtils functions}{

The RobsRUtils functions ...
   \itemize{
     \item \code{\link{ErrorBar_LL_UL}}
     \item \code{\link{FormatJSON_String}}
     \item \code{\link{GetColourBlindPalette}}
     \item \code{\link{GetColourBlindPalette}}
     \item \code{\link{GetPaletteSize}}
     \item \code{\link{MD5_thisFile}}
     \item \code{\link{MakePie}}
     \item \code{\link{OutputPlotForPaper}}
     \item \code{\link{ProgressDot}}
     \item \code{\link{PlotPDF}}
     \item \code{\link{ProducePlot}}
     \item \code{\link{SHA1_thisFile}}
     \item \code{\link{SHA256_thisFile}}
     \item \code{\link{TrimLeading}}
     \item \code{\link{TrimTrailing}}
     \item \code{\link{Trim}}
   }
}

