% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{SHA1_thisFile}
\alias{SHA1_thisFile}
\title{Utility to return SHA1 signature of a file}
\usage{
SHA1_thisFile(filename, writeSHA1File = TRUE)
}
\arguments{
\item{writeSHA1File}{= TRUE - will write a file with the same
base name and extension .sha1 to the same dir as the file}
}
\value{
an uppercase SHA1 signature
}
\description{
Utility to return SHA1 signature of a file
}
\examples{
\dontrun{
SHA1_thisFile(filename)
}

}

