% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{SHA256_thisFile}
\alias{SHA256_thisFile}
\title{Utility to return SHA256 signature of a file}
\usage{
SHA256_thisFile(filename, writeSHA256File = TRUE)
}
\arguments{
\item{writeSHA256File}{= TRUE - will write a file with the same
base name and extension .sha256 to the same dir as the file}
}
\value{
an uppercase SHA256 signature
}
\description{
Utility to return SHA256 signature of a file
}
\examples{
\dontrun{
SHA256_thisFile(filename)
}

}
