% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{Trim}
\alias{Trim}
\title{Utility to return string w/o leading or trailing whitespace}
\usage{
Trim(x)
}
\arguments{
\item{x}{- the string to process}
}
\value{
string w/o trailing whitespace
}
\description{
Utility to return string w/o leading or trailing whitespace
}
\examples{
\dontrun{
filename<-'  File_1234.json   '
filename.clean<-Trim(filename)
}
}
