% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{TrimLeading}
\alias{TrimLeading}
\title{Utility to return string w/o leading whitespace}
\usage{
TrimLeading(x)
}
\arguments{
\item{x}{- the string to process}
}
\value{
string w/o leading whitespace
}
\description{
Utility to return string w/o leading whitespace
}
\examples{
\dontrun{
filename<-'  File_1234.json'
filename.clean<-TrimLeading(filename)
}
}
