% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{TrimTrailing}
\alias{TrimTrailing}
\title{Utility to return string w/o trailing whitespace}
\usage{
TrimTrailing(x)
}
\arguments{
\item{x}{- the string to process}
}
\value{
string w/o trailing whitespace
}
\description{
Utility to return string w/o trailing whitespace
}
\examples{
\dontrun{
filename<-'File_1234.json   '
filename.clean<-TrimTrailing(filename)
}
}
