% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobsRUtils.R
\name{VerifyMD5}
\alias{VerifyMD5}
\title{Wrapper function for checking an MD5}
\usage{
VerifyMD5(full.path, MD5, stop.program = TRUE)
}
\arguments{
\item{MD5}{- the expected MD5}

\item{filename}{- the filename for the file to check}

\item{stop.program=TRUE}{- what happens if it fails}
}
\description{
Wrapper function for checking an MD5
}
\examples{
\dontrun{

data.dir <- './data'
data.set.filename <- 'A-File-To-Check.csv'
full.path <- paste0(data.dir,'/',data.set.filename)

VerifyMD5(full.path,'3363CFFB842AB858F425D5E3DC546BF4')

}
}
